Add files via upload
Hangman is a popular word guessing game where the player attempts to build a missing word by guessing one letter at a time. The player is provided with certain number of chances and has to guess wisely the letter with the total number of letters in that word being the only hint provided to him. The game ends if a certain number of chances are exploited and the word is still incomplete or fully missing. The game also gets over when the whole word has been guessed correctly by the player. Traditionally, chances are tracked using a stick figure drawing of a person being hanged from a gallows. The figure is drawn one body part at a time, and the guesser loses when the entire figure has been drawn.